Move Clipper into a separate module
